:sparkles: Make API asynchronous
Since timeouts can occur when calling WLED (especially in the test function) it blocks things, until the timeouts have been resolved. Using backoff doesn't help here, but it is useful to help with random connectivity errors so it stays.

Issue was known, but became an actual problem with the initial GET request blocking UI startup as no other requests could be made.